+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React, { useRef, useState } from "react";
-import Nickname from "./Nickname";
-import Toggle from "./Toggle";
+import Nickname from "./components/Nickname";
+import Toggle from "./components/Toggle";
 import { User } from "./model/User";
 import Id from "./Id";
 import { ToggleData } from "@/src/model/ToggleData";
