+++ src/components/Nickname/style.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const NicknameContainer = styled.div``;
+
+export const NicknameItem = styled.div`
+  background-color: #0daedd;
+  color: white;
+  margin: 10px;
+  border-radius: 5px;
+  cursor: pointer;
+`;
