+++ src/components/Toggle/index.tsx
@@ -1,26 +1,18 @@
+import { ToggleProps } from "@/src/types/toggle";
 import React from "react";
 import { Form } from "react-bootstrap";
-
-interface ToggleProps {
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  label: string;
-  value: boolean;
-}
+import { ToggleItem } from "./style";
 
 const Toggle = ({ onChange, label, value }: ToggleProps) => {
   return (
-    <li
-      style={{
-        listStyleType: "none",
-      }}
-    >
+    <ToggleItem>
       <Form.Check
         onChange={onChange}
         type="switch"
         label={label}
         checked={value}
       />
-    </li>
+    </ToggleItem>
   );
 };
