+++ src/components/Toggle/style.ts
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const ToggleItem = styled.li`
+  list-style: none;
+`;
