+++ src/types/toggle.ts
@@ -1,0 +1,5 @@
+export interface ToggleProps {
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  label: string;
+  value: boolean;
+}
